chore: reorganize controllers
